flask_app/requirements.txt
python-dotenv
flask
requests
streamlit
beautifulsoup4
tiktoken==0.6.0
flask-cors==4.0.0